rename folder, use hook instead of EnrollmentButton
