--- airbyte-webapp/src/packages/cloud/components/experiments/FreeConnectorProgram/EnrollmentButton/index.ts
@@ -1,1 +1,0 @@
-export { FCPEnrollmentButton } from "./EnrollmentButton";
