+++ airbyte-webapp/src/packages/cloud/components/experiments/FreeConnectorProgram/EnrollmentModal/FreeSvg.tsx
