+++ airbyte-webapp/src/packages/cloud/components/experiments/FreeConnectorProgram/EnrollmentModal/MailSvg.tsx
